SiRI_Recomentations/requirements.txt
asgiref==3.6.0
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
joblib==1.2.0
nltk==3.8.1
numpy==1.24.3
psycopg2-binary==2.9.6
pycparser==2.21
pycryptodome==3.17
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2023.3
regex==2023.5.5
sqlparse==0.4.3
tqdm==4.65.0
